feat: created token folder, move all token files to that folder. Refactored add token page
