+++ packages/ui/src/components/token/TokenSearchView.tsx
@@ -1,0 +1,211 @@
+import AutoSizer from "react-virtualized-auto-sizer"
+import { FixedSizeList as List } from "react-window"
+
+// Components
+import TokenDisplay from "./TokenDisplay"
+import { useOnMountHistory } from "../../context/hooks/useOnMount"
+
+// Assets
+import searchIcon from "../../assets/images/icons/search.svg"
+import { TokenResponse } from "../../routes/settings/AddTokensPage"
+import { useState } from "react"
+import PopupFooter from "../popup/PopupFooter"
+import { ButtonWithLoading } from "../button/ButtonWithLoading"
+
+const SearchedTokenView = (props: any) => {
+    const history = useOnMountHistory()
+    const [message, setMessage] = useState<string>("")
+    const [selected, setSelected] = useState<TokenResponse[]>([])
+
+    const { isSearchEmpty, results, searchedValue } = props
+
+    // Handlers
+    const onSubmit = async () => {
+        try {
+            // Valid form data
+            if (selected.length > 0) {
+                history.push({
+                    pathname: "/settings/tokens/add/confirm",
+                    state: {
+                        tokens: selected,
+                        searchedValue: searchedValue,
+                        ...(history.location.state || {}),
+                    },
+                })
+            } else {
+                // Prevent manual form submission
+                setMessage("Please select a token first.")
+            }
+        } catch (event) {
+            // Invalid form data
+            console.log(event)
+        }
+    }
+
+    // Functions
+    const addToken = (token: TokenResponse) => {
+        setSelected(selected.concat(token))
+    }
+
+    const removeToken = (token: TokenResponse) => {
+        setSelected(selected.filter((el) => el.address !== token.address))
+    }
+
+    const onClick = (token: TokenResponse) => {
+        // Check if the token is already selected
+        if (!selected.some((el) => el.address === token.address)) {
+            // Add selected token
+            addToken(token)
+
+            // Reset message
+            if (message !== "") {
+                setMessage("")
+            }
+        } else {
+            // Remove selected token
+            removeToken(token)
+        }
+    }
+
+    const filteredResults = results.filter(
+        (result: { address: string }) =>
+            !selected.some((el) => el.address === result.address)
+    )
+
+    return (
+        <div className="flex flex-col justify-between w-full h-full">
+            <div className="w-full h-full mt-16">
+                {/* ERROR */}
+                <div
+                    className={`text-xs px-6 text-red-500 ${
+                        message === "" ? "pt-0 h-0" : "pt-2"
+                    }`}
+                >
+                    {message || <>&nbsp;</>}
+                </div>
+                {/* HINT */}
+                {isSearchEmpty && selected.length <= 0 ? (
+                    <div className="flex flex-col items-center justify-start flex-1 h-full p-6">
+                        <div className="flex justify-center items-center relative mb-6">
+                            <img
+                                src={searchIcon}
+                                alt="search"
+                                className="w-7 h-7 absolute z-10"
+                            />
+                            <div className="w-20 h-20 bg-primary-100 rounded-full relative z-0"></div>
+                        </div>
+                        <span className="text-sm text-gray-600 text-center">
+                            Add the tokens that you've acquired using
+                            BlockWallet.
+                            <br />
+                            Enter an address for adding a custom token.
+                        </span>
+                    </div>
+                ) : (
+                    <div className="flex-1 flex flex-col w-full h-0 max-h-screen px-6 pb-0">
+                        <div
+                            className={`text-xs text-gray-500 pt-4 pb-0 ${
+                                selected.length <= 0 ? "hidden" : "visible"
+                            }`}
+                        >
+                            SELECTED TOKENS
+                        </div>
+                        <div className="flex flex-col">
+                            {selected.map((select) => {
+                                // Selected tokens
+                                return (
+                                    <div
+                                        className="cursor-pointer"
+                                        key={`selected-${select.address}`}
+                                        onClick={() => onClick(select)}
+                                    >
+                                        <TokenDisplay
+                                            data={select}
+                                            clickable={false}
+                                            active={true}
+                                            hoverable={true}
+                                            textSize="sm"
+                                        />
+                                    </div>
+                                )
+                            })}
+                        </div>
+                        <div
+                            className={`text-xs text-gray-500 pt-4 pb-1 ${
+                                isSearchEmpty ? "hidden" : "visible"
+                            }`}
+                        >
+                            SEARCH TOKENS
+                        </div>
+                        <div className="flex flex-col">
+                            {results.length < 1 && selected.length <= 0 ? (
+                                <div className="text-base font-bold text-black w-full text-center mt-4">
+                                    No match
+                                </div>
+                            ) : (
+                                <div
+                                    style={{
+                                        height: 314,
+                                    }}
+                                    className="w-full"
+                                >
+                                    <AutoSizer>
+                                        {({ width, height }) => (
+                                            <List
+                                                height={height}
+                                                width={width}
+                                                itemCount={
+                                                    filteredResults.length
+                                                }
+                                                itemSize={60}
+                                                itemData={filteredResults}
+                                            >
+                                                {({ style, data, index }) => (
+                                                    <div
+                                                        style={style}
+                                                        className="cursor-pointer"
+                                                        key={`result-${data[index].address}`}
+                                                        onClick={() =>
+                                                            onClick(data[index])
+                                                        }
+                                                    >
+                                                        <TokenDisplay
+                                                            data={data[index]}
+                                                            clickable={false}
+                                                            active={false}
+                                                            hoverable={true}
+                                                            textSize="sm"
+                                                        />
+                                                    </div>
+                                                )}
+                                            </List>
+                                        )}
+                                    </AutoSizer>
+                                </div>
+                            )}
+                        </div>
+
+                        {/* FOOTER */}
+                        {!(results.length < 1 && selected.length <= 0) ? (
+                            <div className="w-full">
+                                <hr className="border-0.5 border-gray-200 w-full" />
+                                <PopupFooter>
+                                    <ButtonWithLoading
+                                        label="Next"
+                                        formId="search-form"
+                                        disabled={selected.length === 0}
+                                        onClick={onSubmit}
+                                    />
+                                </PopupFooter>
+                            </div>
+                        ) : (
+                            <></>
+                        )}
+                    </div>
+                )}
+            </div>
+        </div>
+    )
+}
+
+export default SearchedTokenView
